+++ changelog.md
@@ -21,6 +21,9 @@ Codion Change Log
 ### is.codion.framework.db.rmi
 ### is.codion.framework.db.http
 - javadocs updated, some improvements and fixes.
+### is.codion.framework.json.domain
+### is.codion.framework.server
+- javadocs updated, some improvements and fixes.
 ### is.codion.framework.model
 - AbstractEntityEditor.execute() and supersede() protected again.
 - DefaultEntityModel, DefaultEntityEditModel and DefaultEntityApplicationModel renamed AbstractEntityModel, AbstractEntityEditModel and AbstractEntityApplicationModel, now abstract, consistent with the other Abstract* extension bases.